+++ app/javascript/vue/tasks/collecting_events/new_collecting_event/store/actions/processGeoreferenceQueue.js
@@ -1,12 +1,14 @@
-import { CreateGeoreference } from '../../request/resources'
+import { CreateGeoreference, UpdateGeoreference } from '../../request/resources'
 import { MutationNames } from '../mutations/mutations'
 
 export default async ({ state, commit }) => {
   const promises = []
 
   state.queueGeoreferences.forEach(georeference => {
+    const saveGeoreference = georeference.id ? UpdateGeoreference : CreateGeoreference
+
     georeference.collecting_event_id = state.collectingEvent.id
-    promises.push(CreateGeoreference(georeference).then(response => {
+    promises.push(saveGeoreference(georeference).then(response => {
       commit(MutationNames.AddGeoreference, response.body)
     }))
   })