+++ CHANGELOG.md
@@ -28,6 +28,7 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 - Refresh button component [#2085]
 - Update comprehensive url [#2096]
 - `/units.json` called 2x [#2089]
+- Edit `error radius` of a georeference in new collecting event task [#2087]
 
 ### Changed
 - Now using ImageMagick 7 instead of 6
@@ -45,6 +46,7 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 [#2066]: https://github.com/SpeciesFileGroup/taxonworks/issues/2066
 [#2071]: https://github.com/SpeciesFileGroup/taxonworks/issues/2071
 [#2085]: https://github.com/SpeciesFileGroup/taxonworks/issues/2085
+[#2087]: https://github.com/SpeciesFileGroup/taxonworks/issues/2087
 [#2089]: https://github.com/SpeciesFileGroup/taxonworks/issues/2089
 [#2094]: https://github.com/SpeciesFileGroup/taxonworks/issues/2094
 [#2096]: https://github.com/SpeciesFileGroup/taxonworks/issues/2096